revert launch arg modifications
