+++ crates/cubecl-core/src/frontend/element/cube_elem.rs
@@ -1,16 +1,13 @@
-use crate::compute::KernelLauncher;
 use crate::frontend::UInt;
 use crate::frontend::{CubeType, ExpandElement};
 use crate::ir::{Elem, Variable};
-use crate::Runtime;
 
-use super::{ArgSettings, ExpandElementTyped, LaunchArg, LaunchArgExpand, Vectorized};
+use super::{ExpandElementTyped, Vectorized};
 
 /// Form of CubeType that encapsulates all primitive types:
 /// Numeric, UInt, Bool
 pub trait CubePrimitive:
     CubeType<ExpandType = ExpandElementTyped<Self>>
-    + LaunchArgExpand
     + Vectorized
     + core::cmp::Eq
     + core::cmp::PartialEq
@@ -20,8 +17,6 @@ pub trait CubePrimitive:
     + Clone
     + Copy
 {
-    type Primitive: ScalarArgSettings;
-
     /// Return the element type to use on GPU
     fn as_elem() -> Elem;
 
@@ -40,10 +35,6 @@ macro_rules! impl_into_expand_element {
     };
 }
 
-impl<T: CubePrimitive> LaunchArg for T {
-    type RuntimeArg<'a, R: Runtime> = ScalarArg<T>;
-}
-
 impl_into_expand_element!(u32);
 impl_into_expand_element!(usize);
 impl_into_expand_element!(bool);
@@ -57,23 +48,5 @@ impl From<UInt> for ExpandElement {
         ExpandElement::Plain(crate::ir::Variable::ConstantScalar(
             crate::ir::ConstantScalarValue::UInt(value.val as u64),
         ))
-    }
-}
-
-/// Similar to [ArgSettings], however only for scalar types that don't depend on the [Runtime]
-/// trait.
-pub trait ScalarArgSettings: Send + Sync {
-    /// Register the information to the [KernelLauncher].
-    fn register<R: Runtime>(&self, launcher: &mut KernelLauncher<R>);
-}
-
-#[derive(new)]
-pub struct ScalarArg<T: CubePrimitive> {
-    elem: T::Primitive,
-}
-
-impl<T: CubePrimitive, R: Runtime> ArgSettings<R> for ScalarArg<T> {
-    fn register(&self, launcher: &mut crate::compute::KernelLauncher<R>) {
-        self.elem.register(launcher);
     }
 }
