+++ crates/cubecl-core/src/frontend/element/float.rs
@@ -77,7 +77,6 @@ macro_rules! impl_float {
         }
 
         impl CubePrimitive for $type {
-            type Primitive = $primitive;
             /// Return the element type to use on GPU
             fn as_elem() -> Elem {
                 Elem::Float(FloatKind::$type)
@@ -104,7 +103,9 @@ macro_rules! impl_float {
             }
         }
 
-        impl Numeric for $type {}
+        impl Numeric for $type {
+            type Primitive = $primitive;
+        }
 
         impl From<u32> for $type {
             fn from(val: u32) -> Self {
