+++ crates/cubecl-core/src/frontend/element/int.rs
@@ -58,7 +58,6 @@ macro_rules! impl_int {
         }
 
         impl CubePrimitive for $type {
-            type Primitive = $primitive;
             fn as_elem() -> Elem {
                 Elem::Int(IntKind::$type)
             }
@@ -102,7 +101,9 @@ macro_rules! impl_int {
             }
         }
 
-        impl Numeric for $type {}
+        impl Numeric for $type {
+            type Primitive = $primitive;
+        }
 
         impl ExpandElementBaseInit for $type {
             fn init_elem(context: &mut CubeContext, elem: ExpandElement) -> ExpandElement {
