+++ crates/cubecl-core/src/frontend/element/numeric.rs
@@ -1,12 +1,17 @@
+use crate::compute::KernelLauncher;
 use crate::frontend::{CubeContext, CubePrimitive, CubeType};
 use crate::ir::{Item, Variable};
 use crate::prelude::Clamp;
+use crate::Runtime;
 use crate::{
     frontend::{index_assign, Abs, Max, Min, Remainder},
     unexpanded,
 };
 
-use super::{ExpandElement, ExpandElementBaseInit, ExpandElementTyped, LaunchArgExpand, UInt, I64};
+use super::{
+    ArgSettings, ExpandElement, ExpandElementBaseInit, ExpandElementTyped, LaunchArg,
+    LaunchArgExpand, UInt, I64,
+};
 
 /// Type that encompasses both (unsigned or signed) integers and floats
 /// Used in kernels that should work for both.
@@ -45,6 +50,8 @@ pub trait Numeric:
     + std::cmp::PartialOrd<u32>
     + std::cmp::PartialEq<u32>
 {
+    type Primitive: ScalarArgSettings;
+
     /// Create a new constant numeric.
     ///
     /// Note: since this must work for both integer and float
@@ -92,4 +99,26 @@ pub trait Numeric:
 
         new_var.into()
     }
+}
+
+/// Similar to [ArgSettings], however only for scalar types that don't depend on the [Runtime]
+/// trait.
+pub trait ScalarArgSettings: Send + Sync {
+    /// Register the information to the [KernelLauncher].
+    fn register<R: Runtime>(&self, launcher: &mut KernelLauncher<R>);
+}
+
+#[derive(new)]
+pub struct ScalarArg<T: Numeric> {
+    elem: T::Primitive,
+}
+
+impl<T: Numeric, R: Runtime> ArgSettings<R> for ScalarArg<T> {
+    fn register(&self, launcher: &mut crate::compute::KernelLauncher<R>) {
+        self.elem.register(launcher);
+    }
+}
+
+impl<T: Numeric> LaunchArg for T {
+    type RuntimeArg<'a, R: Runtime> = ScalarArg<T>;
 }
