+++ crates/cubecl-core/src/frontend/element/uint.rs
@@ -38,7 +38,6 @@ impl ExpandElementBaseInit for UInt {
 }
 
 impl CubePrimitive for UInt {
-    type Primitive = u32;
     fn as_elem() -> Elem {
         Elem::UInt
     }
@@ -60,7 +59,9 @@ impl ScalarArgSettings for u32 {
     }
 }
 
-impl Numeric for UInt {}
+impl Numeric for UInt {
+    type Primitive = u32;
+}
 
 impl UInt {
     pub const fn new(val: u32) -> Self {
